LayoutTests/media/media-source/media-source-real-overlapping-dts.html

<!DOCTYPE html>
<html>
<head>
    <title>media-source-real-overlapping-dts</title>
    <script src="mp4-generator.js"></script>
    <script src="../video-test.js"></script>
    <script>
    // Tests appending with composition time offsets (PTS != DTS) using real MP4 data.
    // This verifies that the real parser correctly handles samples where
    // presentation order differs from decode order (B-frame-like patterns).

    var source;
    var sourceBuffer;

    async function runTest() {
        findMediaElement();

        if (!MediaSource.isTypeSupported(MP4.VIDEO_TYPE)) {
            failTest(`${MP4.VIDEO_TYPE} is not supported`);
            return;
        }

        source = new MediaSource();
        run('video.src = URL.createObjectURL(source)');
        await waitFor(source, 'sourceopen');

        run(`sourceBuffer = source.addSourceBuffer('${MP4.VIDEO_TYPE}')`);

        var initData = MP4.initSegment({
            timescale: 1000,
            tracks: [{ id: 1, type: 'video', timescale: 1000 }]
        });
        sourceBuffer.appendBuffer(initData);
        await waitFor(sourceBuffer, 'updateend');

        // Samples with PTS != DTS (B-frame-like decode ordering):
        // DTS order: 0, 1000, 2000, 3000
        // PTS order: 0, 2000, 1000, 3000 (swap samples 1 and 2)
        // CTO = PTS - DTS: 0, 1000, -1000, 0
        var media = MP4.mediaSegment({
            sequenceNumber: 1,
            tracks: [{
                id: 1,
                type: 'video',
                baseDecodeTime: 0,
                samples: [
                    { duration: 1000, compositionTimeOffset: 0, isSync: true },      // DTS=0, PTS=0
                    { duration: 1000, compositionTimeOffset: 1000, isSync: false },   // DTS=1, PTS=2
                    { duration: 1000, compositionTimeOffset: -1000, isSync: false },  // DTS=2, PTS=1
                    { duration: 1000, compositionTimeOffset: 0, isSync: false },      // DTS=3, PTS=3
                ]
            }]
        });
        sourceBuffer.appendBuffer(media);
        await waitFor(sourceBuffer, 'updateend');

        // Buffered range should cover PTS 0 to 4 (since all samples are present)
        testExpected('sourceBuffer.buffered.length', 1);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.start(0)', 0, 0.01);
        testExpectedEqualWithTolerance('sourceBuffer.buffered.end(0)', 4, 0.01);

        endTest();
    }

    </script>
</head>
<body onload="runTest()">
    <div>Tests B-frame-like composition time offsets (PTS != DTS) with real MP4 data.</div>
    <video></video>
</body>
</html>
