LayoutTests/media/media-source/media-source-real-overlapping-dts-expected.txt

Tests B-frame-like composition time offsets (PTS != DTS) with real MP4 data.

RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer('video/mp4; codecs="avc1.4d401e"'))
EVENT(updateend)
EVENT(updateend)
EXPECTED (sourceBuffer.buffered.length == '1') OK
EXPECTED (sourceBuffer.buffered.start(0) == '0') OK
EXPECTED (sourceBuffer.buffered.end(0) == '4') OK
END OF TEST
